Separate Logout component from login; stop setting username state on Login screen to close #202
